experiments/requirements.txt
anytree==2.8.0
appdirs==1.4.4
attrs==21.2.0
bcrypt==3.2.0
Beaker==1.11.0
btrfsutil==5.14.2
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.6
chardet==4.0.0
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==35.0.0
cycler==0.10.0
distlib==0.3.3
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
ds4drv==0.5.1
evdev==1.4.0
greenlet==1.1.0
html5lib==1.1
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.3.2
lensfun==0.3.95
louis==3.19.0
lxml==4.6.3
Mako==1.1.5
mallard-ducktype==1.0.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
mne==0.23.4
more-itertools==8.10.0
msgpack==1.0.2
neovim==0.3.1
numpy==1.21.2
ordered-set==4.0.2
packaging==20.9
pandas==1.3.3
paramiko==2.7.2
pep517==0.11.0
Pillow==8.3.2
ply==3.11
progress==1.6
pyarrow==6.0.0
pycairo==1.20.1
pycparser==2.20
pygame==2.0.2
Pygments==2.10.0
PyGObject==3.40.1
PyNaCl==1.4.0
pynvim==0.4.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytoml==0.1.21
pytz==2021.1
pyudev==0.22.0.dev20201112
PyWavelets==1.2.0
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
scikit-learn==1.0.1
scipy==1.7.1
SCons==4.2.0
six==1.16.0
sklearn==0.0
sounddevice==0.4.3
team==1.0
texttable==1.6.4
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
urllib3==1.26.7
webencodings==0.5.1
websocket-client==0.59.0
zipp==3.6.0
